move `getRecipeFluidInputAmount()` to MultiblockFuelRecipeLogic
